calculates the cardim model predictions

Calculates the cardim model predictions by using the predict_males() and predict_females() functions. The data.frame must have the following structure.

Variable	Type

Id	numeric or string

sex	"Male" or "Female"

Age, years	numeric

Income less than \(18000\)€	dichotomic

Diabetes duration, years	numeric

hba1c, \(\%\)	numeric

Hypertension treatment at baseline	dichotomic

Logarithm of albumine creatinine ratio, log(mg/g)	numeric

\(\text{Total cholesterol} - \text{hdl cholesterol}\), mmol/L	numeric

Physical activity, inactive	dichotomic

Physical activity, partially active	dichotomic

Atrial fibrillation at baseline	dichotomic

\(\text{Systolic blood pressure} - \text{diastolic blood pressure}\), mmHg	numeric

Retinopathy at baseline	dichotomic

Smoking status, smoker	dichotomic

Smoking status, ex-smoker	dichotomic

predict(data, na_mean = TRUE)

Arguments

data

data.frame or tibble where all the patients data is stored

na_mean

substitute the missing values with NA (default: TRUE)

Value

cardim predictions for the given cohort

Examples

predict(
  tibble::tibble(
  id = c(1,2,3),
  sex = c("Male", "Female", "Female"),
  age = c(60, 65, 45),
  diabetes_duration = c(6, 10, 2),
  hba1c = c(9.8, 6.7, 8.5),
  hypertension_treatment = c(1, 1, 0),
  income_less_18000 = c(1, 1, 1),
  log_albtocreatratio = c(2.25, 1.84, 2.36),
  non_hdl = c(3.95, 4.3, 3.4),
  pulse_pressure = c(70, 74, 65),
  physical_activity_inactive = c(0, 0, 0),
  physical_activity_partially_active = c(1, 0, 1),
  previous_atrial_f = c(0, 0, 0),
  retinopathy = c(1, 1, 0),
  smoking_status_smoker = c(0, 0, 1),
  smoking_status_ex_smoker = c(0, 0, 0)
  )
)
#> # A tibble: 3 × 17
#>      id sex      age diabetes_duration hba1c hypertension_treatment
#>   <dbl> <chr>  <dbl>             <dbl> <dbl>                  <dbl>
#> 1     1 Male      60                 6   9.8                      1
#> 2     2 Female    65                10   6.7                      1
#> 3     3 Female    45                 2   8.5                      0
#> # ℹ 11 more variables: income_less_18000 <dbl>, log_albtocreatratio <dbl>,
#> #   non_hdl <dbl>, pulse_pressure <dbl>, physical_activity_inactive <dbl>,
#> #   physical_activity_partially_active <dbl>, previous_atrial_f <dbl>,
#> #   retinopathy <dbl>, smoking_status_smoker <dbl>,
#> #   smoking_status_ex_smoker <dbl>, pred <dbl>