calculates the cardim males model prediction

calculates the model prediction for the given parameters. The risk predictions are calculated as follows.

$$ \text{prediction} = 1 - S_0(t)^{exp(\beta X)} $$

where \(\beta\) is the model coefficients and \(S_0(t)\) is \(1 - \lambda_0(t)\) that is the cumulative incidence at the time \(t\).

predict_males(
  age,
  diabetes_duration,
  hba1c,
  hypertension_treatment,
  log_albtocreatratio,
  non_hdl,
  income_less_18000,
  physical_activity_inactive,
  physical_activity_partially_active,
  previous_atrial_f,
  pulse_pressure,
  retinopathy,
  smoking_status_smoker,
  smoking_status_ex_smoker,
  na_mean = TRUE,
  .test_variables = TRUE
)

Arguments

age

Age in years

diabetes_duration

Diabetes duration in years

hba1c

HbA1c in %

hypertension_treatment

if the patient has an active hypertension treatment at baseline, 1 = yes, 0 = no

log_albtocreatratio

logarithm of the albumine creatinine ratio in log(mg/g)

non_hdl

\(\text{total cholesterol} - \text{hdl cholesterol}\) in mmol/L

income_less_18000

if the patient is in the category `income_less_18000`, 1 = yes, 0 = no

physical_activity_inactive

if the patient is in the category `inactive`, 1 = yes, 0 = no

physical_activity_partially_active

if the patient is in the category `partially_inactive`, 1 = yes , 0 = no

previous_atrial_f

if the patient had atrial fibrillation at baseline, 1 = yes, 0 = no

pulse_pressure

\(\text{systolic blood pressure} - \text{diastolic blood pressure}\) in mmHg

retinopathy

if the patient had retinopathy at baseline, 1 = yes, 0 = no

smoking_status_smoker

if the patient is in the category `smoker`, 1 = yes, 0 = no

smoking_status_ex_smoker

if the patient is in the category `ex-smoker`, 1 = yes, 0 = no

na_mean

substitute the missing values with `NA` (default: `TRUE`)

.test_variables

if the function should test the arguments type. (could lead to errors)

Value

model prediction for the given parameters

Examples

predict_males(
  age = 71,
  diabetes_duration = 24,
  hba1c = 9.9,
  hypertension_treatment = 1,
  log_albtocreatratio = 4.577,
  non_hdl = 2.793,
  income_less_18000 = 1,
  physical_activity_inactive = 0,
  physical_activity_partially_active = 0,
  previous_atrial_f = 0,
  pulse_pressure = 67,
  retinopathy = 1,
  smoking_status_smoker = 0,
  smoking_status_ex_smoker = 0
)
#> [1] 0.2673847